big aileron overhaul
does each axis independently, and uses existing gcode points closest to
where it should be cutting (good enough?)

it will error out if something between the two profiles is different
enough to make them come out different lengths.  cross this bridge when
I come to it.

also only does ailerons that are a fixed % of the wing chord, so it
tapers as the wing tapers.  no idea if this makes sense, but it puts it
at the same place, aerodynamically, everywhere, which seems to be what
the DLGs do...