working toward apogee-style wing
there's a bug in the merging profiles routine that causes the two AG
airfoils to come out with slightly different lengths, even though they
are the same in the files -- with different coordinates

is probably a bug.

workaround is to test for equal length, and if so, skip the profile
merge step.  Will work for apogee, at least.
Might fall down on aileron wings if the spacings are large.