more fixes -- tested against all wings
and working!  so that's good.

now converting to selig format

a problem I was having was that files started/ended with 1,0
so that data point would appear twice
I manually edited all the data files, but that's stupid
instead, should account for it while reading data in

and there's still some glitch with e297.dat, that is probably worth
debugging